lLinting
Former-commit-id: b427eeaf3d4f9d62b239296615ceb2279d786bd4 [formerly 772c617578a23770d952d71f1e542790e1fb2a83]
Former-commit-id: 059bf9f072ae62658498e81a3d67bd55e3b3520c

+++ api/models/categories/find_test.go
@@ -49,7 +49,7 @@ func (t *CategoriesTestSuite) TestStore_Find() {
 	for name, test := range tt {
 		t.Run(name, func() {
 			s := t.Setup(test.mock)
-			got, err := s.Find(int64(category.Id))
+			got, err := s.Find(category.Id)
 			if err != nil {
 				t.Contains(errors.Message(err), test.want)
 				return
